refactor: make it WORK
